+++ Changes.md
@@ -1,7 +1,10 @@
 1.6.x.x (relative to 1.6.17.0)
 =======
 
+Fixes
+-----
 
+- RenderMan : Fixed interactive edits to volume transforms.
 
 1.6.17.0 (relative to 1.6.16.0)
 ========